App => Navigation Improvement  | v0.2
