+++ App/App.js
@@ -1,10 +1,8 @@
 import { StatusBar } from 'expo-status-bar';
-import { StyleSheet, Text, View } from 'react-native';
 import { NavigationContainer } from '@react-navigation/native';
-import { createMaterialTopTabNavigator } from '@react-navigation/material-top-tabs';
+import { MyTabs } from './src/Main';
+import Constants from 'expo-constants';
 
-
-const Tab = createMaterialTopTabNavigator();
 
 export default function App() {
   return (
@@ -13,45 +11,10 @@ export default function App() {
         barStyle='light-content'
         backgroundColor='#000'
       />
-     <MyTabs/>
+      <MyTabs style={{ marginTop: Constants.statusBarHeight }} />
     </NavigationContainer>
   );
 }
 
-function HomeScreen() {
-  return (
-    <View style={styles.container}>
-      <Text style={styles.text}>Hello World!</Text>
-    </View>
-  )
-}
 
-function Lol() {
-  return (
-    <View style={styles.container}>
-      <Text style={styles.text}> </Text>
-    </View>
-  )
-}
 
-function MyTabs() {
-  return (
-    <Tab.Navigator>
-      <Tab.Screen name="Home" component={HomeScreen} />
-      <Tab.Screen name="Settings" component={Lol} />
-    </Tab.Navigator>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#000',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  text: {
-    color: '#fff',
-    fontSize: 50
-  }
-});
