+++ App/src/routes/HomeScreen.jsx
@@ -1,0 +1,22 @@
+import { StyleSheet, Text, View } from 'react-native';
+
+export function HomeScreen() {
+    return (
+        <View style={styles.container}>
+            <Text style={styles.text}>Hello World!</Text>
+        </View>
+    )
+}
+
+const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        backgroundColor: '#000',
+        alignItems: 'center',
+        justifyContent: 'center',
+    },
+    text: {
+        color: '#fff',
+        fontSize: 50
+    }
+});
